v0.8.8 - Fix filtering of data before batch processing
Previously filtered out records that had `record.id === undefined` which led to unintended consequences. Some record data is just an array of IDs, others includes `field` info but no id (e.g. 'create' record data).

This change reverts to the previous state where all data is assumed to be valid before passing to `batchProcess`. If an error occurs, it should be up to the caller to catch and handle.